+++ Forms/Administrative/AdministrativeSubject.py
@@ -1,8 +1,6 @@
 import sys
-from AdministrativeAccess import *
-from Validaciones import *
-from loadSubjects import *
-from loadStudents import *
+from Validations.loadSubjects import *
+from Validations.loadStudents import *
 from tkinter import *
 from tkinter import ttk
 from tkinter import messagebox
@@ -103,5 +101,5 @@ class AdministrativeSubject:
 
     def returnAdministrativeHome(self):
         ##Add validations to return or close and open the other window
-        window = __import__('AdministrativeAccess')
+        window = __import__('Forms.Administrative.AdministrativeAccess',None,None,['AdministrativeAccess'], 0)
         self.app = window.AdministrativeAccess(self.new_root,self.clave)
