+++ Main.py
@@ -1,23 +1,6 @@
-import sys
-from StudentSchedule import *
-from GroupsInscription import *
-from Acceso import *
+from Forms.Access import * 
 from tkinter import *
-from tkinter import ttk
 
 if __name__ == '__main__':
     """ CREACIÓN DE LA VENTANA DE ACCESO AL SISTEMA"""
-
-    #Declara ventana de aplicación
-
-
-    #LLamado a la aplicación mediante la ventana de acceso al sistema
-    #Access(Tk())
-
-    #Inscription(Tk(),2)# prueba de nueva ventana
-
-    StudentSchedule(Tk(),2)
-
-    GroupsInscription(Tk(),2,1)
-    #Bucle de la aplicación
-#root.mainloop()
+    Access(Tk())
